个人代码备份

+++ cloud-common/cloud-auto-coder/src/main/resources/cloud-springfox/serviceImpl.java.ftl
@@ -2,9 +2,9 @@ package ${package}.service.impl;
 
 import com.baomidou.mybatisplus.core.conditions.query.LambdaQueryWrapper;
 import com.baomidou.mybatisplus.core.metadata.IPage;
-import com.cloud.common.core.utils.BeanUtil;
-import com.cloud.common.mybatis.page.PageParam;
-import com.cloud.common.mybatis.util.OrderUtil;
+import com.tianque.scgrid.userprecision.service.utils.BeanUtil;
+import com.tianque.doraemon.mybatis.support.Condition;
+import com.tianque.doraemon.mybatis.support.PageParam;
 import ${package}.entity.${table.className};
 import ${package}.vo.${table.className}Vo;
 import ${package}.mapper.${table.className}Mapper;
@@ -22,6 +22,7 @@ import ${basePackage}${foreignKey.packagePath}.service.${foreignKey.tableNameCla
 import org.springframework.stereotype.Service;
 import org.springframework.util.CollectionUtils;
 
+import java.util.ArrayList;
 import java.util.Collections;
 import java.util.List;
 <#if table.foreignKeys?? && (table.foreignKeys?size > 0) >
@@ -30,16 +31,24 @@ import java.util.stream.Collectors;
 </#if>
 
 /**
+ *  ${table.comment}
+ *
  * @author liulei
  */
 @Service
 public class ${table.className}ServiceImpl implements ${table.className}Service {
 
+    /**
+     * ${table.comment}Mapper
+     */
     private ${table.className}Mapper ${table.className? uncap_first}Mapper;
 
 <#if table.foreignKeys?? && (table.foreignKeys?size > 0) >
     <#list table.foreignKeys as foreignKey>
 <#if foreignKey.tableNameClass != table.className>
+    /**
+     * ${foreignKey.comment}服务
+     */
     private ${foreignKey.tableNameClass}Service ${foreignKey.tableNameClass? uncap_first}Service;
 
 </#if>
@@ -69,7 +78,7 @@ public class ${table.className}ServiceImpl implements ${table.className}Service 
     }
 
     /**
-     * 保存对象
+     * 保存${table.comment}
      *
      * @param ${table.className? uncap_first}Param 前端传入对象
      * @return 返回保存成功状态
@@ -85,7 +94,7 @@ public class ${table.className}ServiceImpl implements ${table.className}Service 
     }
 
     /**
-     * 通过Id查询数据
+     * 通过Id查询${table.comment}数据
      *
      * @param id 业务Id
      * @return 返回VO对象
@@ -100,25 +109,30 @@ public class ${table.className}ServiceImpl implements ${table.className}Service 
     }
 
     /**
-     * 传入多个Id 查询数据
+     * 传入多个Id 查询${table.comment}数据
      *
      * @param ids 多个id
      * @return  返回list结果
      */
     @Override
     public List<${table.className}Vo> findByIds(List<Long> ids) {
+        if(CollectionUtils.isEmpty(ids)) {
+            new ArrayList();
+        }
         LambdaQueryWrapper<${table.className}> queryWrapper = new LambdaQueryWrapper<>();
         queryWrapper.in(${table.className}::getId, ids);
         List<${table.className}> ${table.className? uncap_first}Entities = ${table.className? uncap_first}Mapper.selectList(queryWrapper);
+        //数据转换
         List<${table.className}Vo> list = BeanUtil.copyListProperties(${table.className? uncap_first}Entities, ${table.className}Vo::new);
 		<#if table.foreignKeys?? && (table.foreignKeys?size > 0) >
+        //封装关联数据
 		this.setParam(list);
 		</#if>
 		return list;
     }
 
     /**
-     * 根据查询条件 查询列表
+     * 根据查询条件 查询${table.comment}列表
      *
      * @param ${table.className? uncap_first}Query 查询条件
      * @return 返回list结果
@@ -127,6 +141,7 @@ public class ${table.className}ServiceImpl implements ${table.className}Service 
     public List<${table.className}Vo> findByList(${table.className}Query ${table.className? uncap_first}Query) {
         IPage<${table.className}Vo> iPage = this.queryPage(${table.className? uncap_first}Query, new PageParam());
 		<#if table.foreignKeys?? && (table.foreignKeys?size > 0) >
+        //封装关联数据
 		this.setParam(iPage.getRecords());
 		</#if>
         return iPage.getRecords();
@@ -140,31 +155,35 @@ public class ${table.className}ServiceImpl implements ${table.className}Service 
      */
     @Override
     public Boolean removeByIds(List<Long> ids) {
-        LambdaQueryWrapper<${table.className}> queryWrapper = new LambdaQueryWrapper<>();
-        queryWrapper.in(${table.className}::getId, ids);
-        ${table.className? uncap_first}Mapper.delete(queryWrapper);
+        if(!CollectionUtils.isEmpty(ids)) {
+            LambdaQueryWrapper<${table.className}> queryWrapper = new LambdaQueryWrapper<>();
+            queryWrapper.in(${table.className}::getId, ids);
+            ${table.className? uncap_first}Mapper.delete(queryWrapper);
+        }
         return Boolean.TRUE;
     }
 
     /**
      * 数据分页查询
      *
-     * @param ${table.className? uncap_first}Query 查询条件
+     * @param ${table.className? uncap_first}Query  查询条件
      * @param pageParam 分页条件
      * @return 分页数据
      */
     @Override
     public IPage<${table.className}Vo> queryPage(${table.className}Query ${table.className? uncap_first}Query, PageParam pageParam) {
         IPage<${table.className}> iPage = ${table.className? uncap_first}Mapper.queryPage(OrderUtil.getPage(pageParam), ${table.className? uncap_first}Query);
+        //数据转换
         IPage<${table.className}Vo> page = iPage.convert(${table.className? uncap_first} -> BeanUtil.copyProperties(${table.className? uncap_first}, ${table.className}Vo::new));
 		<#if table.foreignKeys?? && (table.foreignKeys?size > 0) >
+        //封装关联数据
 		this.setParam(page.getRecords());
 		</#if>
         return page;
     }
 
     /**
-     * 通过Id 更新数据
+     * 通过Id 更新${table.comment}
      *
      * @param ${table.className? uncap_first} 前端更新集合
      * @return  更新成功状态
@@ -180,15 +199,19 @@ public class ${table.className}ServiceImpl implements ${table.className}Service 
 	<#list table.keys as key>
 	
 	/**
-     * 传入多个Id 查询数据
+     * 传入多个${key.tableComment}业务Id 查询数据
      * @param ${key.columnNameClass? uncap_first}s
      * @return
      */
     @Override
     public List<${table.className}Vo> findBy${key.columnNameClass}(List<Long> ${key.columnNameClass? uncap_first}s){
+        if (${key.columnNameClass? uncap_first}s == null || ${key.columnNameClass? uncap_first}s.size() <= 0) {
+            return new ArrayList<>();
+        }
 		LambdaQueryWrapper<${table.className}> queryWrapper = new LambdaQueryWrapper<>();
         queryWrapper.in(${table.className}::get${key.columnNameClass}, ${key.columnNameClass? uncap_first}s);
         List<${table.className}> ${table.className? uncap_first}Entities = ${table.className? uncap_first}Mapper.selectList(queryWrapper);
+        //数据转换
         List<${table.className}Vo> list = BeanUtil.copyListProperties(${table.className? uncap_first}Entities, ${table.className}Vo::new);
         return list;
 	}
@@ -202,19 +225,22 @@ public class ${table.className}ServiceImpl implements ${table.className}Service 
      * @param list 列表数据
      */
     private void setParam(List<${table.className}Vo> list) {
-        if (list != null) {
+        if (!CollectionUtils.isEmpty(list)) {
             List<Long> ids = list.stream().map(${table.className}Vo::getId).collect(Collectors.toList());
 	<#list table.foreignKeys as foreignKey>
+            //查询 ${foreignKey.comment} 数据
             Map<Long, List<${foreignKey.tableNameClass}Vo>> ${foreignKey.tableNameClass? uncap_first}Map = <#if foreignKey.tableNameClass != table.className>${foreignKey.tableNameClass? uncap_first}Service.</#if>findBy${foreignKey.columnNameClass}(ids).stream().collect(Collectors.groupingBy(${foreignKey.tableNameClass}Vo::get${foreignKey.columnNameClass}));
 	</#list>
             for (${table.className}Vo ${table.className? uncap_first} : list) {
 	<#list table.foreignKeys as foreignKey>
         <#if foreignKey.uni>
+                //设置 ${foreignKey.comment} 数据
                 List<${foreignKey.tableNameClass}Vo> ${foreignKey.tableNameClass? uncap_first}Vos = ${foreignKey.tableNameClass? uncap_first}Map.get(${table.className? uncap_first}.getId());
                 if (${foreignKey.tableNameClass? uncap_first}Vos != null && ${foreignKey.tableNameClass? uncap_first}Vos.size() > 0) {
                     ${table.className? uncap_first}.set${foreignKey.tableNameClass}VO(${foreignKey.tableNameClass? uncap_first}Vos.get(0));
                 }
         <#else>
+                //设置 ${foreignKey.comment} 数据
                 ${table.className? uncap_first}.set${foreignKey.tableNameClass}VOList(${foreignKey.tableNameClass? uncap_first}Map.get(${table.className? uncap_first}.getId()));
         </#if>
     </#list>
